+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.106.51](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.160.17...@standardnotes/desktop@3.106.51) (2023-05-30)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.106.50](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.160.16...@standardnotes/desktop@3.106.50) (2023-05-29)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
 ## [3.106.49](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.160.15...@standardnotes/desktop@3.106.49) (2023-05-26)
 
 **Note:** Version bump only for package @standardnotes/desktop